Changement de fichier pour la création de la variable f2op
